hackathon-team/requirements.txt
# Core LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0

# LLM providers
openai>=1.50.0
anthropic>=0.34.0

# Web search (no API key needed)
duckduckgo-search>=6.0.0

# Vector DB for RAG (local, no key needed)
chromadb>=0.5.0

# Web framework (Plumber's pipeline)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database + state
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# UI
streamlit>=1.39.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
httpx>=0.27.0
tenacity>=8.5.0
